requirements.txt
Flask==3.0.3
pandas==2.2.2
numpy==1.26.4
great-expectations==1.3.5
matplotlib==3.9.2
seaborn==0.13.2
openpyxl==3.1.5
xlrd==2.0.1
python-dotenv==1.0.1
Werkzeug==3.0.3